APP-64: Change "add token" text and default screen in follow token
